requirements.txt
black==22.1.0
Django>=3.2.12,<4.0
djangorestframework>=3.13,<4.0
faker-vehicle==0.2.0
flake8==4.0.1
isort==5.10.1
psycopg2>2.9,<3.0
requests>2.27,<3.0
